code update after review
